refactor(spanner): restructure client API exports and modules
Restructures the `google-cloud-spanner` crate exports to present a
modular, clean API design. Removes all legacy re-exports from the
`client` module to establish distinct, dedicated modules.

Module structure:
- Crate root: Re-exports core clients (`Spanner`, `DatabaseClient`),
  statements, mutations, rows, result sets, values, and basic types.
- `client` module: Exposes only `Spanner`, `DatabaseClient`, and
  `DatabaseAdminClient`.
- `transaction` module: Consolidates transaction-scoped builders,
  retry policies, options, and runners.
- `builder` module: Consolidates all client/request builders.
- `model` module: Exposes gapic dataplane model objects.

Rationale:
- Ensures a clear division between transaction-level structures
  and core connection/data-access structures.
- Eliminates duplicate import pathways (e.g., importing types from both
  root and `client`).
- Standardizes the Spanner crate with conventions established in
  workspace crates like Storage and Pub/Sub.
- Cleans up and aligns internal documentation, samples, and all 120
  doctests to strictly use these new import paths.
- Types that need to be used by any application that wants to do anything
  useful with this client are exported at the root level.

Fixes #5430
