+++ src/spanner/examples/src/mutation/update_data.rs
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 // [START spanner_update_data]
-use google_cloud_spanner::client::{DatabaseClient, Mutation};
+use google_cloud_spanner::Mutation;
+use google_cloud_spanner::client::DatabaseClient;
 
 pub async fn sample(client: &DatabaseClient) -> anyhow::Result<()> {
     let mutations = vec![
