+++ src/spanner/examples/src/query/query_data.rs
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 // [START spanner_query_data]
-use google_cloud_spanner::client::{DatabaseClient, Statement};
+use google_cloud_spanner::Statement;
+use google_cloud_spanner::client::DatabaseClient;
 
 pub async fn sample(client: &DatabaseClient) -> anyhow::Result<()> {
     let statement = Statement::builder("SELECT SingerId, AlbumId, AlbumTitle FROM Albums").build();
