+++ src/spanner/src/batch.rs
@@ -1,0 +1,19 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Re-exports for Spanner batch execution and query partitioning.
+
+pub use crate::batch_dml::BatchDml;
+pub use crate::batch_read_only_transaction::{BatchReadOnlyTransaction, Partition};
+pub use crate::batch_write_transaction::{BatchWriteResponseStream, BatchWriteTransaction};
