+++ src/spanner/src/batch_write_transaction.rs
@@ -59,7 +59,7 @@ impl BatchWriteTransaction {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Mutation, Spanner, MutationGroup};
+    /// # use google_cloud_spanner::{Mutation, Spanner, MutationGroup};
     /// # use google_cloud_gax::error::rpc::Code;
     /// # async fn sample() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
@@ -146,7 +146,8 @@ impl BatchWriteResponseStream {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::client::{Mutation, Spanner};
+    use crate::Mutation;
+    use crate::client::Spanner;
     use crate::result_set::tests::adapt;
     use anyhow::Result;
     use gaxi::grpc::tonic::Response;
