+++ src/spanner/src/builder.rs
@@ -1,0 +1,32 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Re-exports for Spanner client and request builders.
+
+pub use crate::batch_dml::BatchDmlBuilder;
+pub use crate::batch_read_only_transaction::BatchReadOnlyTransactionBuilder;
+pub use crate::batch_write_transaction::BatchWriteTransactionBuilder;
+pub use crate::client::ClientBuilder as SpannerBuilder;
+pub use crate::database_client::DatabaseClientBuilder;
+pub use crate::key::KeySetBuilder;
+pub use crate::mutation::WriteBuilder;
+pub use crate::partitioned_dml_transaction::PartitionedDmlTransactionBuilder;
+pub use crate::read::ConfiguredReadRequestBuilder;
+pub use crate::read::ReadRequestBuilder;
+pub use crate::read_only_transaction::MultiUseReadOnlyTransactionBuilder;
+pub use crate::read_only_transaction::SingleUseReadOnlyTransactionBuilder;
+pub use crate::transaction_runner::TransactionRunnerBuilder;
+pub use crate::write_only_transaction::WriteOnlyTransactionBuilder;
+
+pub use google_cloud_spanner_admin_database_v1::builder::database_admin::ClientBuilder as DatabaseAdminBuilder;
