+++ src/spanner/src/database_client.rs
@@ -58,7 +58,8 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Statement;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let tx = db_client.single_use().build();
@@ -81,7 +82,8 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Statement;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let tx = db_client.read_only_transaction().build().await?;
@@ -104,7 +106,8 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Statement;
     /// # async fn build(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let transaction = db_client.batch_read_only_transaction().build().await?;
@@ -122,7 +125,8 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Statement;
     /// # async fn run(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let transaction = db_client.partitioned_dml_transaction().build().await?;
@@ -145,7 +149,8 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Statement;
     /// # async fn build(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let runner = db_client.read_write_transaction().build().await?;
@@ -169,7 +174,8 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```rust
-    /// # use google_cloud_spanner::client::{Mutation, Spanner};
+    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Mutation;
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
     /// let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -199,7 +205,7 @@ impl DatabaseClient {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Mutation, MutationGroup};
+    /// # use google_cloud_spanner::{Spanner, Mutation, MutationGroup};
     /// # use google_cloud_gax::error::rpc::Code;
     /// # async fn sample() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
