+++ src/spanner/src/row.rs
@@ -81,7 +81,7 @@ impl Row {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::{Spanner, Statement};
     /// # async fn test_doc() -> anyhow::Result<()> {
     /// let client = Spanner::builder().build().await?;
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -113,7 +113,7 @@ impl Row {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::{Spanner, Statement};
     /// # async fn test_doc() -> anyhow::Result<()> {
     /// let client = Spanner::builder().build().await?;
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -141,7 +141,7 @@ impl Row {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::{Spanner, Statement};
     /// # async fn test_doc() -> anyhow::Result<()> {
     /// let client = Spanner::builder().build().await?;
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -184,7 +184,7 @@ impl Row {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Spanner, Statement};
+    /// # use google_cloud_spanner::{Spanner, Statement};
     /// # async fn test_doc() -> anyhow::Result<()> {
     /// let client = Spanner::builder().build().await?;
     /// let db_client = client.database_client("projects/p/instances/i/databases/d").build().await?;
