+++ src/spanner/src/transaction.rs
@@ -1,0 +1,24 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Re-exports for Spanner transactions.
+
+pub use crate::partitioned_dml_transaction::PartitionedDmlTransaction;
+pub use crate::read_only_transaction::{
+    BeginTransactionOption, MultiUseReadOnlyTransaction, SingleUseReadOnlyTransaction,
+};
+pub use crate::read_write_transaction::ReadWriteTransaction;
+pub use crate::transaction_retry_policy::BasicTransactionRetryPolicy;
+pub use crate::transaction_runner::TransactionRunner;
+pub use crate::write_only_transaction::WriteOnlyTransaction;
