+++ src/spanner/src/write_only_transaction.rs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::client::{DatabaseClient, Mutation, amend_request_options_for_lar};
+use crate::Mutation;
+use crate::client::{DatabaseClient, amend_request_options_for_lar};
 use crate::model::request_options::Priority;
 use crate::model::transaction_options::ReadWrite;
 use crate::model::{
@@ -100,7 +101,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Spanner;
     /// # use wkt::Duration;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -125,7 +126,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Spanner;
     /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     /// let transaction = db_client.write_only_transaction()
@@ -151,7 +152,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Mutation, Spanner};
+    /// # use google_cloud_spanner::{Mutation, Spanner};
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// # let client = Spanner::builder().build().await?;
     /// # let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -183,7 +184,8 @@ impl WriteOnlyTransactionBuilder {
     /// # Example
     /// ```
     /// # use std::time::Duration;
-    /// # use google_cloud_spanner::client::{BasicTransactionRetryPolicy, Spanner};
+    /// # use google_cloud_spanner::Spanner;
+    /// # use google_cloud_spanner::transaction::BasicTransactionRetryPolicy;
     /// # async fn build_tx(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
     ///
@@ -210,7 +212,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Spanner;
     /// # use std::time::Duration;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -229,7 +231,7 @@ impl WriteOnlyTransactionBuilder {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::Spanner;
+    /// # use google_cloud_spanner::Spanner;
     /// # use google_cloud_gax::retry_policy::NeverRetry;
     /// # async fn sample(spanner: Spanner) -> Result<(), google_cloud_spanner::Error> {
     /// let db_client = spanner.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -369,7 +371,7 @@ impl WriteOnlyTransaction {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Mutation, Spanner};
+    /// # use google_cloud_spanner::{Mutation, Spanner};
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
     /// let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
@@ -490,7 +492,7 @@ impl WriteOnlyTransaction {
     ///
     /// # Example
     /// ```
-    /// # use google_cloud_spanner::client::{Mutation, Spanner};
+    /// # use google_cloud_spanner::{Mutation, Spanner};
     /// # async fn test_doc() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = Spanner::builder().build().await?;
     /// let db = client.database_client("projects/p/instances/i/databases/d").build().await?;
