+++ src/spanner/tests/execute_query.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use google_cloud_spanner::Statement;
 use google_cloud_spanner::client::Spanner;
-use google_cloud_spanner::client::Statement;
 use google_cloud_test_macros::tokio_test_no_panics;
 use spanner_grpc_mock::MockSpanner;
 use spanner_grpc_mock::google::spanner::v1 as mock_v1;
