+++ tests/spanner/src/batch_dml.rs
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use google_cloud_spanner::BatchUpdateError;
-use google_cloud_spanner::batch_dml::BatchDml;
-use google_cloud_spanner::client::{DatabaseClient, Mutation, Statement};
+use google_cloud_spanner::batch::BatchDml;
+use google_cloud_spanner::client::DatabaseClient;
+use google_cloud_spanner::error::BatchUpdateError;
+use google_cloud_spanner::{Mutation, Statement};
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 
 pub async fn successful_batch_update(db_client: &DatabaseClient) -> anyhow::Result<()> {
