+++ tests/spanner/src/directed_read.rs
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use google_cloud_spanner::client::{DatabaseClient, KeySet, ReadRequest};
+use google_cloud_spanner::client::DatabaseClient;
 use google_cloud_spanner::model::DirectedReadOptions;
 use google_cloud_spanner::model::directed_read_options::{IncludeReplicas, Replicas};
+use google_cloud_spanner::{KeySet, ReadRequest};
 
 pub async fn read_only_with_directed_read(db_client: &DatabaseClient) -> anyhow::Result<()> {
     let mut dro = DirectedReadOptions::default();
