+++ tests/spanner/src/partitioned_dml.rs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use google_cloud_spanner::client::{DatabaseClient, Mutation, Statement};
+use google_cloud_spanner::client::DatabaseClient;
+use google_cloud_spanner::{Mutation, Statement};
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 
 pub async fn partitioned_dml_update(db_client: &DatabaseClient) -> anyhow::Result<()> {
