+++ tests/spanner/src/pg_dialect.rs
@@ -14,7 +14,8 @@
 
 use anyhow::Result;
 use google_cloud_spanner::Decimal;
-use google_cloud_spanner::client::{DatabaseClient, Mutation, Statement};
+use google_cloud_spanner::client::DatabaseClient;
+use google_cloud_spanner::{Mutation, Statement};
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 use std::str::FromStr;
