+++ tests/spanner/src/read.rs
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use google_cloud_spanner::client::{DatabaseClient, KeyRange, KeySet, Mutation, ReadRequest};
+use google_cloud_spanner::client::DatabaseClient;
 use google_cloud_spanner::key;
+use google_cloud_spanner::{KeyRange, KeySet, Mutation, ReadRequest};
 use google_cloud_test_utils::resource_names::LowercaseAlphanumeric;
 
 pub async fn read_single_key(db_client: &DatabaseClient) -> anyhow::Result<()> {
